refactor: reports dir
