+++ src/payload.config.ts
@@ -15,7 +15,7 @@ import { Notifications } from './collections/notificaitons'
 import { Pages } from './collections/Pages'
 import { Posts } from './collections/Posts'
 import { ReadNotification } from './collections/ReadNotification'
-import { Reports } from './collections/Reports'
+import { Reports } from './collections/reports/Reports'
 import { Roles } from './collections/Roles'
 import { Settings } from './collections/Settings'
 import { Students } from './collections/students/Students'
